pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [{ name = "Marco André Mendes", email = "marcoandre@gmail.com" }]
dependencies = [
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.3.0",
    "django-extensions>=3.2.3",
    "django-filter>=23.3",
    "drf-spectacular>=0.26.5",
    "gunicorn>=21.2.0",
    "netifaces>=0.11.0",
    "pre-commit>=3.5.0",
    "psycopg2-binary>=2.9.9",
    "pydotplus>=2.0.2",
    "python-dotenv>=1.0.0",
    "setuptools>=69.0.2",
    "whitenoise[brotli]>=6.6.0",
    "python-magic>=0.4.27",
    "Pillow>=10.1.0",
]
requires-python = ">=3.10"
license = { text = "MIT" }

[tool.pdm]

[tool.pdm.dev-dependencies]
format = ["isort>=5.11.3", "black>=22.12.0"]
dev = [
    "isort>=5.12.0",
    "black>=23.10.1",
    "pylint>=3.0.2",
]

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"
createsuperuser = "python manage.py createsuperuser"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
post_migrate = "python manage.py graph_models -g -o core.png core"
shell = "python manage.py shell"
shellp= "python manage.py shell_plus"
test = "python manage.py test"
loaddata = "python manage.py loaddata core.json"
dumpdata = "python manage.py dumpdata --indent 2"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
  \.git
 | \.hg
 | \.mypy_cache
 | \.tox
 | \.venv
 | _build
 | buck-out
 | build
 | __pypackages__
 )/
'''

